added new welcome

+++ frontend/src/actions/stockActions.js
@@ -33,7 +33,7 @@ export const getStocks = () => dispatch => {
 
 
 export const handleChange = (e) => dispatch=>{
-  // console.log(e.target.name)
+
   e.target.name ==='username'?
   dispatch(
     { type: 'CHANGE_USER', 'username': e.target.value}
@@ -45,13 +45,6 @@ export const handleChange = (e) => dispatch=>{
   
 }
 
-// **export const popup =()=> dispatch=> {
-//       <Popup trigger={<button> Trigger </button>} position="right center">
-//         <div>
-//           Popup here
-//         </div>
-//       </Popup>
-//     }
 
 export const getPortfolio =(e)=> dispatch => {
   // e.preventDefault()
@@ -78,7 +71,6 @@ export const getPortfolio =(e)=> dispatch => {
 export const handleSignUp=(e)=> dispatch => {
     
   e.preventDefault()
-  // console.log(e.target.firstname.value)
 
   let newUserObject ={
     firstname: e.target.firstname.value,
@@ -161,9 +153,6 @@ export const handleSignUp=(e)=> dispatch => {
       localStorage.setItem('invested_balance', data.user.invested_balance)
       
       dispatch({type:'LOGIN_SUCCESS', user: data.user, logged: true, portfolio: data.user.portfolios})
-      // debugger
-     
-
     }
   })
 
@@ -176,11 +165,13 @@ export const buyStock = (e, eachStock, balance)=> dispatch=> {
   // console.log(e.target)
   
   let stock ={
+
     price: parseFloat(eachStock['close']),
     ticker: eachStock['ticker'],
     quantity: parseFloat(e.target.quantity.value),
     total_price: parseFloat(eachStock['close'] * e.target.quantity.value),
     user_id: parseFloat(localStorage.uid)
+
   }
 
 if(balance.remaining_balance < stock.total_price){
@@ -259,19 +250,16 @@ export const  sellStock = (e, eachStock,balance)=> dispatch=> {
   let stock ={
     price: parseFloat(eachStock.price),
     ticker: eachStock.ticker,
-    quantity: -parseFloat(e.target.quantity.value),
+    quantity: -parseInt(e.target.quantity.value),
     total_price: -parseFloat(eachStock.price * e.target.quantity.value),
-    user_id: parseFloat(localStorage.uid)
+    user_id: parseInt(localStorage.uid)
 
   }
-
 
   let user ={
     remaining_balance: balance.remaining_balance - stock.total_price,
     invested_balance: balance.invested_balance + stock.total_price
   }
-  
-
   
   fetch(`http://localhost:3000/portfolios`,{
     method: "POST",
